test(storage-resize-images): update tests
